Minor cleanup after spec tests
I'm satisfying my OCD by:

* Wrapping comments
* Removing unused `curses` import
* Changing consts to CONSTS
* Swapping get_meta_schema for a session-cached fixture
* Removing some unnecessary `alias` stuff
* Switching if/elif to match/case
